backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg
alembic
google-generativeai
httpx
firebase-admin
google-cloud-firestore
pydantic
python-multipart
